requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
mysql-connector-python==9.4.0
playwright==1.55.0
propcache==0.4.1
pyee==13.0.0
python-dotenv==1.0.0
pytz==2024.1
redis[asyncio]==5.0.1
requests==2.32.5
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
zipp==3.23.0
